+++ core/buildBuild.ts
@@ -1,7 +1,10 @@
 /// <reference types="bun" />
 
+const virginEngineVersion = JSON.stringify((await Bun.file(`./package.json`).json()).version);
+
 // Build build
 await build();
+await updateVirginEngineVersionHtml();
 
 async function build() {
   const htmlTemplate = minifyHtml(await Bun.file(`./core/build/template.html`).text());
@@ -19,7 +22,7 @@ async function build() {
     target: `bun`,
     define: {
       REPLACE_HTML_TEMPLATE: JSON.stringify(htmlTemplate),
-      REPLACE_VIRGINE_ENGINE_VERSION: JSON.stringify((await Bun.file(`./package.json`).json()).version),
+      REPLACE_VIRGINE_ENGINE_VERSION: virginEngineVersion,
       // REPLACE_CORE: core, // do not works, why?? it copyies only first line!
     },
   });
@@ -64,4 +67,16 @@ function minifyHtml(text: string) {
     .replaceAll(/ " | "|" /g, `"`)
     .replaceAll(/ , | ,|, /g, `,`)
     .replaceAll(`: `, `:`); // color: red; => color:red;
+}
+
+async function updateVirginEngineVersionHtml() {
+  const htmlFile = Bun.file(`./src/app/index.html`);
+  let htmlText = await htmlFile.text();
+
+  htmlText = htmlText.replace(
+    /<title>.*?<\/title>/i,
+    `<title>VirginEngine v${virginEngineVersion.slice(1, -1)}</title>`,
+  );
+
+  await htmlFile.write(htmlText);
 }